Split up estimator and probability distribution
