+++ src/openfermion/measurements/__init__.py
@@ -19,8 +19,9 @@ from ._rdm_equality_constraints import (one_body_fermion_constraints,
                                         two_body_fermion_constraints)
 
 from ._bayesian_estimators import (BayesDepolarizingEstimator,
-                                   BayesEstimator,
-                                   FourierProbabilityDist,)
+                                   BayesEstimator)
+
+from ._probability_distributions import (FourierProbabilityDist)
 
 from ._data_containers import (QPEExperimentData,
                                QPERoundData)
